fastapi_project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
websockets==12.0
bcrypt==4.0.1
python-multipart==0.0.6